+++ SourceCode/Yuconz/src/Employee.java
@@ -3,27 +3,28 @@
  * stores the details of an employee
  *
  * @author Alice Jaffray and Kieran D'Arcy
- * @version 2019/02/15
+ * @version 2019/02/28
  */
 public class Employee extends User {
+
     /**
      * constructor
-     *
-     * @param username the username of the employee
-     * @param password the password of the employee
+     * @param empNo employee's unique employee number.
+     * @param manager employees direct manager
+     * @param department department the employee works for.
      */
-    public Employee(String username, String password) {
-        super(username, password, "employee");
+    public Employee(String empNo, String manager, String department) {
+        super(empNo, department, AccessLevel.EMPLOYEE, manager);
     }
 
     /**
-     * second constructor
-     *
-     * @param username the username of the employee
-     * @param password the password of the employee
-     * @param authlevel the authorisation level of the employee
+     * second constructor for different access levels.
+     * @param empNo user's unique employee number.
+     * @param manager user's direct superior
+     * @param department department the user works in.
+     * @param a access level the user has.
      */
-    public Employee(String username, String password, String authlevel) {
-        super(username, password, authlevel);
+    protected Employee(String empNo, String manager, String department, AccessLevel a) {
+        super(empNo, department, a, manager);
     }
 }