+++ SourceCode/Yuconz/src/Manager.java
@@ -3,16 +3,17 @@
  * Stores the details of a manager
  *
  * @author Alice Jaffray and Kieran D'Arcy
- * @version 2019/02/16
+ * @version 2019/02/28
  */
 public class Manager extends Employee {
+
     /**
      * constructor
-     *
-     * @param username username of the manager
-     * @param password password of the manager
+     * @param empNo unique employee number for this manager.
+     * @param manager direct superior.
+     * @param department department the user works in.
      */
-    public Manager(String username, String password) {
-        super(username, password, "manager");
+    public Manager(String empNo, String manager, String department) {
+        super(empNo, manager, department, AccessLevel.MANAGER);
     }
 }