+++ SourceCode/Yuconz/src/User.java
@@ -3,81 +3,58 @@
  * Stores the details of a user
  *
  * @author Alice Jaffray and Kieran D'Arcy
- * @version 2019/02/15
+ * @version 2019/02/28
  */
 public class User {
-    private String username;
-    private String password;
-    private String authLevel;
-    private String tmpAuthLvl;
+    private String empNo;
+    private AccessLevel accessLevel;
+    private String department;
+    private String manager;
 
     /**
      * constructor
-     *
-     * @param username  the username of the user
-     * @param password  the password of the user
-     * @param authLevel the authorisation level of the user
+     * @param empNo
+     * @param department
+     * @param accessLevel
      */
-    public User(String username, String password, String authLevel) {
-        this.username = username;
-        this.password = password;
-        this.authLevel = authLevel;
-        tmpAuthLvl = authLevel;
+    public User(String empNo, String department, AccessLevel accessLevel, String manager) {
+        this.empNo = empNo;
+        this.department = department;
+        this.accessLevel = accessLevel;
+        this.manager = manager;
     }
 
     /**
-     * gets the username of the user
+     * gets the employee number of the user
      *
-     * @return username the username of the user
+     * @return employee number.
      */
-    public String getUsername() {
-        return username;
+    public String getEmpNo(){
+        return empNo;
     }
 
     /**
-     * gets the password of the user
+     * gets the access level of the user
      *
-     * @return password the password of the user
+     * @return the access level of the user
      */
-    public String getPassword() {
-        return password;
+    public AccessLevel getAccessLevel(){
+        return accessLevel;
     }
 
     /**
-     * gets the authorisation level of the user
-     *
-     * @return authLevel of user of the user
+     * gets user's department.
+     * @return the user's department.
      */
-    public String getAuthLevel() {
-        return authLevel;
+    public String getDepartment() {
+        return department;
     }
 
     /**
-     * sets the authorisation level of the user
-     * prints out the new authorisation level
-     *
-     * @param lvl new wanted authorisation level of user
-     * @return true if changed.
+     * gets the manager of the user.
+     * @return manager.
      */
-    public boolean changeAuthLevel(String lvl) {
-        if(lvl.equals("employee") || lvl.equals("hremployee") || lvl.equals("director") || lvl.equals("manager")) {
-            authLevel = lvl;
-            System.out.println("You now have " + lvl + " privileges.\n" + tmpAuthLvl + " access will be reinstated next time you log in.");
-            return true;
-        } else if(lvl.equals("reset")) {
-            authLevel = tmpAuthLvl;
-            System.out.println("Your Authorisation level has been returned to normal");
-            return true;
-        } else {return false;}
-    }
-
-    /**
-     * gets personal details of a user
-     * returns the details in the form of an array
-     *
-     * @return name and password of user in a neat format
-     */
-    public String getPersonalDetails() {
-        return "Name: " + username + "\nPassword: " + password;
+    public String getManager() {
+        return manager;
     }
 }